+++ lib/custom_social_share.dart
@@ -1,4 +1,4 @@
-library custom_social_share;
+library;
 
 export 'src/custom_social_share.dart';
 export 'src/enums.dart';