+++ lib/src/custom_social_share_method_channel.dart
@@ -19,21 +19,17 @@ class MethodChannelCustomSocialShare extends CustomSocialSharePlatform {
 
   @override
   Future<bool> toAll(String content) {
-    return methodChannel.invokeMethod<bool>('toAll',
-        {"content": content}).then<bool>((bool? value) => value ?? false);
+    return methodChannel.invokeMethod<bool>('toAll', {"content": content}).then<bool>((bool? value) => value ?? false);
   }
 
   @override
   Future<bool> to(ShareWith shareWith, String content) {
-    return methodChannel.invokeMethod<bool>(shareWith.name,
-        {"content": content}).then<bool>((bool? value) => value ?? false);
+    return methodChannel.invokeMethod<bool>(shareWith.name, {"content": content}).then<bool>((bool? value) => value ?? false);
   }
 
   @override
   Future<List<ShareWith>> getInstalledAppsForShare() {
-    return methodChannel
-        .invokeMapMethod<String, bool>('getInstalledApps')
-        .then((map) {
+    return methodChannel.invokeMapMethod<String, bool>('getInstalledApps').then((map) {
       if (map == null) return [];
       map.removeWhere((key, value) => !value);
       return map.keys
@@ -52,9 +48,8 @@ class MethodChannelCustomSocialShare extends CustomSocialSharePlatform {
   @override
   Future<bool> customApp(String package, String content) {
     if (!Platform.isAndroid) return Future.value(false);
-    return methodChannel.invokeMethod<bool>('customApp', {
-      "package": package,
-      "content": content
-    }).then<bool>((bool? value) => value ?? false);
+    return methodChannel
+        .invokeMethod<bool>('customApp', {"package": package, "content": content})
+        .then<bool>((bool? value) => value ?? false);
   }
 }