+++ lib/src/custom_social_share_platform_interface.dart
@@ -42,8 +42,7 @@ abstract class CustomSocialSharePlatform extends PlatformInterface {
 
   /// return list of default [ShareWith] installed apps
   Future<List<ShareWith>> getInstalledAppsForShare() {
-    throw UnimplementedError(
-        'getInstalledAppsForShare() has not been implemented.');
+    throw UnimplementedError('getInstalledAppsForShare() has not been implemented.');
   }
 
   /// custom app share only for Android