+++ test/custom_social_share_test.dart
@@ -5,9 +5,7 @@ import 'package:custom_social_share/src/enums.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
-class MockCustomSocialSharePlatform
-    with MockPlatformInterfaceMixin
-    implements CustomSocialSharePlatform {
+class MockCustomSocialSharePlatform with MockPlatformInterfaceMixin implements CustomSocialSharePlatform {
   @override
   Future<bool> copy(String content) => Future.value(content == 'success');
 
@@ -15,21 +13,18 @@ class MockCustomSocialSharePlatform
   Future<bool> toAll(String content) => Future.value(content == 'success');
 
   @override
-  Future<bool> to(ShareWith shareWith, String content) =>
-      Future.value(ShareWith.sms == shareWith && content == 'success');
+  Future<bool> to(ShareWith shareWith, String content) => Future.value(ShareWith.sms == shareWith && content == 'success');
 
   @override
-  Future<List<ShareWith>> getInstalledAppsForShare() =>
-      Future.value([ShareWith.sms, ShareWith.email]);
+  Future<List<ShareWith>> getInstalledAppsForShare() => Future.value([ShareWith.sms, ShareWith.email]);
 
   @override
-  Future<bool> customApp(String package, String content) => Future.value(
-      package == 'com.google.android.apps.dynamite' && content == 'success');
+  Future<bool> customApp(String package, String content) =>
+      Future.value(package == 'com.google.android.apps.dynamite' && content == 'success');
 }
 
 void main() {
-  final CustomSocialSharePlatform initialPlatform =
-      CustomSocialSharePlatform.instance;
+  final CustomSocialSharePlatform initialPlatform = CustomSocialSharePlatform.instance;
 
   test('$MethodChannelCustomSocialShare is the default instance', () {
     expect(initialPlatform, isInstanceOf<MethodChannelCustomSocialShare>());
@@ -37,8 +32,7 @@ void main() {
 
   test('getPlatformVersion', () async {
     CustomSocialShare customSocialSharePlugin = CustomSocialShare();
-    MockCustomSocialSharePlatform fakePlatform =
-        MockCustomSocialSharePlatform();
+    MockCustomSocialSharePlatform fakePlatform = MockCustomSocialSharePlatform();
     CustomSocialSharePlatform.instance = fakePlatform;
 
     expect(await customSocialSharePlugin.toAll('success'), true);
@@ -47,15 +41,9 @@ void main() {
     expect(await customSocialSharePlugin.to(ShareWith.sms, 'success'), true);
     expect(await customSocialSharePlugin.to(ShareWith.email, 'other'), false);
 
-    expect(await customSocialSharePlugin.getInstalledAppsForShare(),
-        [ShareWith.sms, ShareWith.email]);
+    expect(await customSocialSharePlugin.getInstalledAppsForShare(), [ShareWith.sms, ShareWith.email]);
 
-    expect(
-        await customSocialSharePlugin.customApp(
-            'com.google.android.apps.dynamite', 'success'),
-        true);
-    expect(
-        await customSocialSharePlugin.customApp('com.google.android', 'other'),
-        false);
+    expect(await customSocialSharePlugin.customApp('com.google.android.apps.dynamite', 'success'), true);
+    expect(await customSocialSharePlugin.customApp('com.google.android', 'other'), false);
   });
 }